Fix Cloudwatch Error in Fraud Report (#10874)
* Stub AWS cloudwatch at a lower level to create a failing test for fraud report

* Fix fraud report query that was missing "limit 10000"

changelog: Internal, Reporting, Fix query error in fraud metrics reporting
