+++ spec/rails_helper.rb
@@ -42,6 +42,7 @@ RSpec.configure do |config|
   config.include Capybara::RSpecMatchers, type: :component
   config.include AgreementsHelper
   config.include AnalyticsHelper
+  config.include AwsCloudwatchHelper
   config.include AwsKmsClientHelper
   config.include KeyRotationHelper
   config.include OtpHelper
